# specs/openapi.yaml
openapi: 3.1.0
info:
  title: Seatsurfing REST API
  description: |
    Seatsurfing provides REST APIs used by the mobile and web app as well as the administrator web interface.

    ## Authorization checks

    Each API endpoint is protected by authorization checks:
    1. **Role check**: Role of the user (Anonymous, User, Space Admin, Org Admin, Super Admin)
    2. **Organization check**: Organization membership

    ## Roles

    - **Anonymous**: Every user allowed, no authorization required
    - **User**: Authorized user who is a member of an organization
    - **Space Admin**: An organization's space administrator (can manage locations, spaces, bookings)
    - **Org Admin**: An organization's full administrator
    - **Super Admin**: Global administrator with access to all organizations and all functions

    ## Authentication

    Authentication is done via JWT Bearer tokens in the `Authorization` header. Service accounts may also use HTTP Basic Authentication.

    ## Session management

    - Access Token Lifetime: 15 minutes
    - Refresh Token Lifetime: 28 days
    - Tokens are signed using RS512

    ## Custom error codes

    Some endpoints return custom error codes via the `X-Error-Code` response header:
    - `1001`: Booking slot conflict
    - `1002`: Location max concurrent bookings reached
    - `1003`: Too many upcoming bookings for user
    - `1004`: Too many days in advance
    - `1005`: Invalid booking duration
...
    - `1010`: Booking subject required
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained via login or refresh endpoints
    BasicAuth:
      type: http
      scheme: basic
      description: Service account authentication using `{orgId}_{email}` as username and the service account password
  schemas:
...
    GetLocationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        maxConcurrentBookings:
          type: integer
        timezone:
          type: string
        enabled:
          type: boolean
        mapScale:
          type: number
          format: double
        mapWidth:
          type: integer
        mapHeight:
          type: integer
        mapMimeType:
          type: string
...
    CreateSpaceRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        x:
          type: integer
          description: X position on the location map
        "y":
          type: integer
          description: Y position on the location map
        width:
          type: integer
        height:
          type: integer
        rotation:
          type: integer
        requireSubject:
          type: boolean
        enabled:
          type: boolean
        spaceTypeId:
          type: string
          format: uuid
        attributes:
          type: array
          items:
            $ref: "#/components/schemas/SpaceAttributeValueRequest"
        approverGroupIds:
          type: array
          items:
            type: string
            format: uuid
        allowedBookerGroupIds:
          type: array
          items:
            type: string
            format: uuid
    SpaceAttributeValueRequest:
      type: object
      properties:
        attributeId:
          type: string
          format: uuid
        value:
          type: string
    GetSpaceResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        locationId:
          type: string
          format: uuid
        available:
          type: boolean
        name:
          type: string
        x:
          type: integer
        "y":
          type: integer
        width:
          type: integer
        height:
          type: integer
        rotation:
          type: integer
        requireSubject:
          type: boolean
        enabled:
          type: boolean
        spaceTypeId:
          type: string
          format: uuid
        spaceType:
          $ref: "#/components/schemas/GetSpaceTypeResponse"
        attributes:
          type: array
          items:
            $ref: "#/components/schemas/SpaceAttributeValueRequest"
        approverGroupIds:
          type: array
          items:
            type: string
        allowedBookerGroupIds:
          type: array
          items:
            type: string
        location:
          $ref: "#/components/schemas/GetLocationResponse"
...
    CreateSpaceTypeRequest:
      type: object
      required: [name, bookingMode]
      properties:
        name:
          type: string
          maxLength: 128
        bookingMode:
          type: string
          enum: [flexible_time, fixed_slots]
        minDurationMinutes:
          type: integer
          minimum: 0
        enabled:
          type: boolean
    CreateSpaceTypeSlotRequest:
      type: object
      required: [label, startTime, endTime]
      properties:
        label:
          type: string
          maxLength: 128
        startTime:
          type: string
          pattern: "^\\d{2}:\\d{2}$"
          example: "08:00"
        endTime:
          type: string
          pattern: "^\\d{2}:\\d{2}$"
          example: "12:00"
        enabled:
          type: boolean
        sortOrder:
          type: integer
    GetSpaceTypeSlotResponse:
      allOf:
        - $ref: "#/components/schemas/CreateSpaceTypeSlotRequest"
        - type: object
          properties:
            id:
              type: string
              format: uuid
            spaceTypeId:
              type: string
              format: uuid
    GetSpaceTypeResponse:
      allOf:
        - $ref: "#/components/schemas/CreateSpaceTypeRequest"
        - type: object
          properties:
            id:
              type: string
              format: uuid
            organizationId:
              type: string
              format: uuid
            slots:
              type: array
              items:
                $ref: "#/components/schemas/GetSpaceTypeSlotResponse"
...
  responses:
    Created:
      description: Object created successfully
      headers:
        X-Object-ID:
          schema:
            type: string
          description: ID of the created object
    Updated:
      description: Object updated or deleted successfully (204 No Content)
    BadRequest:
      description: Malformed request or incomplete JSON
      headers:
        X-Error-Code:
          schema:
            type: integer
          description: Custom error code (see error codes table)
...
    Forbidden:
      description: Authorized, but user has no access to requested resource
    NotFound:
      description: Object not found
...
paths:
...
  /location/{locationId}/space/:
    get:
      tags: [Spaces]
      summary: Get all spaces in a location
      description: Returns all spaces in a location with their attributes, approvers, and allowed bookers.
      operationId: getAllSpaces
      security:
        - BearerAuth: []
      parameters:
        - name: locationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of spaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetSpaceResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Spaces]
      summary: Create a space
      description: Creates a new space in a location. Requires Space Admin role.
      operationId: createSpace
      security:
        - BearerAuth: []
      parameters:
        - name: locationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSpaceRequest"
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /location/{locationId}/space/{id}:
    get:
      tags: [Spaces]
      summary: Get a space
      description: Returns a single space by ID with its attributes, approvers, and allowed bookers.
      operationId: getSpace
      security:
        - BearerAuth: []
      parameters:
        - name: locationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Space details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetSpaceResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Spaces]
      summary: Update a space
      description: Updates an existing space. Requires Space Admin role.
      operationId: updateSpace
      security:
        - BearerAuth: []
      parameters:
        - name: locationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSpaceRequest"
      responses:
        "204":
          $ref: "#/components/responses/Updated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [Spaces]
      summary: Delete a space
      description: Deletes a space. Requires Space Admin role.
      operationId: deleteSpace
      security:
        - BearerAuth: []
      parameters:
        - name: locationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          $ref: "#/components/responses/Updated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /space-type/:
    get:
      tags: [Space Types]
      summary: Get all space types
      description: Returns all seat classifications for the user's organization.
      operationId: getAllSpaceTypes
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of space types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetSpaceTypeResponse"
    post:
      tags: [Space Types]
      summary: Create a space type
      description: Creates a new seat classification and booking rule. Requires Space Admin role.
      operationId: createSpaceType
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSpaceTypeRequest"
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /space-type/{id}:
    get:
      tags: [Space Types]
      summary: Get a space type
      description: Returns a single seat classification by ID.
      operationId: getSpaceType
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Space type details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetSpaceTypeResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Space Types]
      summary: Update a space type
      description: Updates an existing seat classification. Requires Space Admin role.
      operationId: updateSpaceType
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSpaceTypeRequest"
      responses:
        "204":
          $ref: "#/components/responses/Updated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [Space Types]
      summary: Delete a space type
      description: Deletes a seat classification if it is not assigned to any spaces. Requires Space Admin role.
      operationId: deleteSpaceType
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          $ref: "#/components/responses/Updated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /space-type/{id}/slot/:
    post:
      tags: [Space Types]
      summary: Create a space type time slot
      description: Adds a fixed booking slot to a seat classification. Requires Space Admin role.
      operationId: createSpaceTypeSlot
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSpaceTypeSlotRequest"
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /space-type/{id}/slot/{slotId}:
    put:
      tags: [Space Types]
      summary: Update a space type time slot
      description: Updates a fixed booking slot. Requires Space Admin role.
      operationId: updateSpaceTypeSlot
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: slotId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSpaceTypeSlotRequest"
      responses:
        "204":
          $ref: "#/components/responses/Updated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Space Types]
      summary: Delete a space type time slot
      description: Deletes a fixed booking slot. Requires Space Admin role.
      operationId: deleteSpaceTypeSlot
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: slotId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          $ref: "#/components/responses/Updated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ===========================
  # Organizations
  # ===========================
...